+++ core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added new render links functionality for map and area ids.
 * Added new option for hiding treasure if all rewards are known.
+* Added new option to show NPC ids to input into other addons.
 
 # v1